+++ src/CPU65816.h
@@ -1,5 +1,5 @@
 // POMIIGS — Apple IIgs emulator
-// Copyright (C) 2026
+// VERHILLE Arnaud — Copyright (C) 2026 — GPLv3 (see LICENSE)
 //
 // This program is free software: you can redistribute it and/or modify it
 // under the terms of the GNU General Public License as published by the Free